refactor: 복잡한 삼항연산자 개선
